requirements.txt
django
djangorestframework
django-filter
django-debug-toolbar
django-eav2
django-mptt
psycopg2-binary
pillow
requests
django-ckeditor
django-rest-knox
celery[redis]
django-environ
gunicorn
drf-yasg
bs4
humanize
django-admin-rangefilter
pydantic[email]
cryptography
djangorestframework-simplejwt
pandas
yookassa
pyTelegramBotAPI
django-cors-headers
python-telegram-bot